+++ internal/resource/url.go
@@ -28,12 +28,14 @@ import (
 	"net/netip"
 	"net/url"
 	"os"
+	"path/filepath"
 	"strings"
 	"sync"
 	"syscall"
 
 	"cloud.google.com/go/compute/metadata"
 	configErrors "github.com/coreos/ignition/v2/config/shared/errors"
+	"github.com/coreos/ignition/v2/internal/distro"
 	"github.com/coreos/ignition/v2/internal/log"
 	"github.com/coreos/ignition/v2/internal/util"
 	"github.com/coreos/vcontext/report"
@@ -60,7 +62,6 @@ const (
 
 var (
 	ErrSchemeUnsupported      = errors.New("unsupported source scheme")
-	ErrPathNotAbsolute        = errors.New("path is not absolute")
 	ErrNotFound               = errors.New("resource not found")
 	ErrFailed                 = errors.New("failed to fetch resource")
 	ErrCompressionUnsupported = errors.New("compression is not supported with that scheme")
@@ -183,6 +184,13 @@ func (f *Fetcher) FetchToBuffer(u url.URL, opts FetchOptions) ([]byte, error) {
 		return buf.Bytes(), err
 	case "gs":
 		err = f.fetchFromGCS(u, dest, opts)
+	case "file":
+		err = f.fetchFromFile(u.Path, dest, opts)
+	case "oem":
+		if !distro.OEMFetch() {
+			return nil, ErrSchemeUnsupported
+		}
+		err = f.fetchFromFile(filepath.Join("/oem", u.Path), dest, opts)
 	case "":
 		return nil, nil
 	default:
@@ -251,6 +259,13 @@ func (f *Fetcher) Fetch(u url.URL, dest *os.File, opts FetchOptions) error {
 		return f.fetchFromS3(u, dest, opts)
 	case "gs":
 		return f.fetchFromGCS(u, dest, opts)
+	case "file":
+		return f.fetchFromFile(u.Path, dest, opts)
+	case "oem":
+		if !distro.OEMFetch() {
+			return ErrSchemeUnsupported
+		}
+		return f.fetchFromFile(filepath.Join("/oem", u.Path), dest, opts)
 	case "":
 		return nil
 	default:
@@ -457,6 +472,18 @@ func (f *Fetcher) fetchFromGCS(u url.URL, dest io.Writer, opts FetchOptions) err
 	}
 
 	return f.fetchFromHTTP(gcsURL, dest, opts)
+}
+
+func (f *Fetcher) fetchFromFile(path string, dest io.Writer, opts FetchOptions) error {
+	path = filepath.Clean(path)
+
+	fi, err := os.Open(path)
+	if err != nil {
+		f.Logger.Err("failed to read file: %v", err)
+		return err
+	}
+	defer func() { _ = fi.Close() }()
+	return f.decompressCopyHashAndVerify(dest, fi, opts)
 }
 
 type s3target interface {